move admin model to src and cleanup
