+++ src/routes/new-user.ts
@@ -1,7 +1,7 @@
 import express, { Request, Response } from 'express';
 import * as Crypto from 'crypto';
 import fs from 'fs';
-import { Admin, AdminDoc } from '../../models/admins';
+import { Admin, AdminDoc } from '../models/admins';
 import { User } from '../models/user';
 import { deflateRaw } from 'zlib';
 import connectEnsureLogin from 'connect-ensure-login';
